packed pricing context

+++ contracts/JBTiered721DelegateStore.sol
@@ -670,12 +670,12 @@ contract JBTiered721DelegateStore is IJBTiered721DelegateStore {
 
       // Add the tier with the iterative ID.
       _storedTierOf[msg.sender][_tierId] = JBStored721Tier({
-        price: uint96(_tierToAdd.price),
+        price: uint104(_tierToAdd.price),
         remainingQuantity: uint32(_tierToAdd.initialQuantity),
         initialQuantity: uint32(_tierToAdd.initialQuantity),
         votingUnits: uint40(_tierToAdd.votingUnits),
         reservedRate: uint16(_tierToAdd.reservedRate),
-        category: uint16(_tierToAdd.category),
+        category: uint24(_tierToAdd.category),
         packedBools: _packBools(
           _tierToAdd.allowManualMint,
           _tierToAdd.transfersPausable,
