+++ contracts/structs/JBStored721Tier.sol
@@ -13,11 +13,11 @@ pragma solidity ^0.8.0;
   @member useVotingUnits A flag indicating if the voting units override should be used over the price as the tier's voting units.
 */
 struct JBStored721Tier {
-  uint96 price;
+  uint104 price;
   uint32 remainingQuantity;
   uint32 initialQuantity;
   uint40 votingUnits;
-  uint16 category;
+  uint24 category;
   uint16 reservedRate;
   uint8 packedBools;
 }
